pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "boxflow"
version = "0.1.0"
description = "AI-assisted image labeling for object detection. Upload, auto-detect, classify, export."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Temur Turayev" }]
keywords = ["labeling", "annotation", "object-detection", "yolo", "clip", "computer-vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.7",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "Pillow>=10.2.0",
    "numpy>=1.26.0",
    "aiofiles>=23.2.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
yolo = ["ultralytics>=8.1.0"]
clip = ["open_clip_torch>=2.26.0"]
all = ["ultralytics>=8.1.0", "open_clip_torch>=2.26.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "httpx>=0.26.0", "ruff>=0.3.0"]

[project.scripts]
boxflow = "boxflow.cli:main"

[project.urls]
Homepage = "https://github.com/TemurTurayev/boxflow"
Repository = "https://github.com/TemurTurayev/boxflow"
Issues = "https://github.com/TemurTurayev/boxflow/issues"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
